refactored register feature
